requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
mysqlclient
pymysql
requests
python-dotenv
gunicorn
PyJWT


